+++ R/utils.R
@@ -99,6 +99,8 @@ format_dates_col <- function(data,
       stop("Dates in dates column must be formatted as dates (YYYY-MM-DD).", call. = FALSE)
     } else {
       data$Date <- as.Date(data$Date, "%Y-%m-%d")
+      if (any(is.na(data$Date)))
+        stop("At least one date in dates column is not a date (YYYY-MM-DD).", call. = FALSE)
     }
   }
   
@@ -146,7 +148,7 @@ format_groups_col <- function(data,
   
   # Rename values to "Value" (and change original if required so no duplication)
   if ("STATION_NUMBER" %in% colnames(data) & groups != "STATION_NUMBER") {
-    names(data)[names(groups) == "STATION_NUMBER"] <- "STATION_NUMBER_orig"
+    names(data)[names(data) == "STATION_NUMBER"] <- "STATION_NUMBER_orig"
   }
   
   names(data)[names(data) == groups] <- "STATION_NUMBER"